refactor: replace Rotaru-Iftene with Mueller modular sqrt algorithm
Replace the precomputation-heavy Rotaru-Iftene implementation with a
stateless Mueller (Cipolla-Lehmer-Mueller) method using Lucas V-sequence
evaluation. The same 2-adic order heuristic decides when to use it, but
no static tables or window-size tuning are needed.

- Rename ModSqrtUtil to ModularSqrt (public entry point: Sqrt -> Compute)
- Remove all precomputation fields, tables, and helper methods
- Add MuellerSqrt with random parameter search and LucasV evaluation
- Simplify InitParams to only store the enableSpeedup flag
- Let TonelliShanks and heuristic fallback obtain the modulus dynamically
