+++ Eduard/Security/Extensions/PointEncodingExtensions.cs
@@ -192,7 +192,7 @@ namespace Eduard.Security.Extensions
                 Array.Copy(bytes, data, bytes.Length - 1);
 
                 BigInteger Yp = new BigInteger(data);
-                BigInteger Xp = ModSqrtUtil.Sqrt(curve.Evaluate(Yp), true);
+                BigInteger Xp = ModularSqrt.Compute(curve.Evaluate(Yp), true);
 
                 int sign = bytes[bytes.Length - 1] - 2;
                 int x_sign = Xp.TestBit(0) ? 1 : 0;
@@ -277,7 +277,7 @@ namespace Eduard.Security.Extensions
                 Array.Copy(bytes, data, bytes.Length - 1);
 
                 BigInteger Xp = new BigInteger(data);
-                BigInteger Yp = ModSqrtUtil.Sqrt(curve.Evaluate(Xp), true);
+                BigInteger Yp = ModularSqrt.Compute(curve.Evaluate(Xp), true);
 
                 int sign = bytes[bytes.Length - 1] - 2;
                 int y_sign = Yp.TestBit(0) ? 1 : 0;
